+++ Products/Reportek/www/static/assign_role.js
@@ -1,0 +1,9 @@
+$(function () {
+  $("#obligations").select2({width: 200});
+  $("#role").select2({
+    width: 200,
+    placeholder: "(All)",
+    allowClear: true
+  });
+  $("#countries, #obligations").select2({width: 200});
+});